criando novo arquivo para manter os metodos
